fix (structure): refactor names of files to match project
