+++ app/src/main/java/com/android/bookswap/SimpleData.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.bookswap
 
 import kotlin.math.sqrt
