+++ app/src/test/java/com/android/bookswap/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.bookswap
 
 import org.junit.Assert.*
 import org.junit.Test
